+++ daml-script/test/src/test-utils/com/daml/lf/engine/script/test/SandboxParticipantFixture.scala
@@ -12,7 +12,7 @@ import org.scalatest.Suite
 import com.daml.bazeltools.BazelRunfiles._
 import com.daml.ledger.api.testing.utils.AkkaBeforeAndAfterAll
 import com.daml.ledger.api.tls.TlsConfiguration
-import com.daml.ledger.runner.common.Config.SandboxParticipantId
+import com.daml.ledger.sandbox.SandboxOnXForTest.SandboxParticipantId
 import com.daml.lf.engine.script.ledgerinteraction.ScriptTimeMode
 
 import scala.concurrent.ExecutionContext
