+++ ledger/sandbox-on-x/src/test/it/scala/com/daml/ledger/platform/sandbox/services/time/WallClockTimeIT.scala
@@ -10,7 +10,7 @@ import com.daml.grpc.GrpcException
 import com.daml.grpc.adapter.client.akka.ClientAdapter
 import com.daml.ledger.api.testing.utils.SuiteResourceManagementAroundAll
 import com.daml.ledger.api.v1.testing.time_service.{GetTimeRequest, SetTimeRequest, TimeServiceGrpc}
-import com.daml.ledger.runner.common.Config.SandboxParticipantId
+import com.daml.ledger.sandbox.SandboxOnXForTest.SandboxParticipantId
 import com.daml.platform.sandbox.fixture.SandboxFixture
 import com.daml.platform.services.time.TimeProviderType
 import org.scalatest.concurrent.{AsyncTimeLimitedTests, ScalaFutures}
