+++ ledger/sandbox-on-x/src/test/lib/scala/com/daml/platform/sandbox/fixture/SandboxFixture.scala
@@ -5,7 +5,7 @@ package com.daml.platform.sandbox.fixture
 
 import com.daml.ledger.api.testing.utils.{OwnedResource, Resource, SuiteResource}
 import com.daml.ledger.resources.{ResourceContext, ResourceOwner}
-import com.daml.ledger.runner.common.Config.SandboxParticipantId
+import com.daml.ledger.sandbox.SandboxOnXForTest.SandboxParticipantId
 import com.daml.ledger.sandbox.SandboxOnXForTest
 import com.daml.platform.apiserver.services.GrpcClientResource
 import com.daml.platform.sandbox.AbstractSandboxFixture
